requirements.txt
apturl==0.5.2
future==1.0.0
asn1crypto==0.24.0
astroid==2.11.7
beautifulsoup4==4.6.0
black==22.8.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==8.0.4
colorama==0.3.7
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
dill==0.3.4
distro==1.9.0
distro-info===0.18ubuntu0.18.04.1
feedparser==5.2.1
graphsurgeon==0.4.5
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==4.8.3
isort==5.10.1
jetson-stats==4.2.7
Jetson.GPIO==2.0.17
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.13.3
oauth==1.0.1
oauthlib==2.0.6
onboard==1.4.1
pandas==0.22.0
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.4.0
protobuf==3.0.0
psutil==5.9.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.5.3
pylint==2.13.9
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.6
python-dateutil==2.9.0.post0
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==6.0.1
requests==2.18.4
requests-unixsocket==0.1.5
RPi.GPIO==0.7.1
scipy==0.19.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
smbus2==0.4.3
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorrt==8.2.1.8
tomli==1.2.3
typed-ast==1.5.5
typing-extensions==4.1.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
wrapt==1.16.0
xkit==0.0.0
youtube-dl==2018.3.14
zipp==3.6.0
zope.interface==4.3.2
